Move PreservationTask entities
- Move `package_.PreservationTask` struct to
  `datatypes.PreservationTask` to avoid circular imports
- Move PreservationTask enums to the `enums` package to avoid circular
  imports
